server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
razorpay==1.4.1
pymongo==4.6.1
python-dateutil==2.8.2
aiosmtplib==3.0.1
